Refactoring MonoBehaviour logic into another class
